Dont flush buffers when no good eviction candidate
This is a bit of a big change, when our user cache is full,
we don't need to flush the buffers to make users "unmodified" anymore
since:
    1. We store trades in redis
    2. We can gather pending orders from the exchange
Thus, a better way to evict users from cache is to remove
those with an "incomplete" view of their pending orders.
I sense this is true because fetching their pending orders
is a pretty expensive operation, so minimizing those calls is best.